Fix validator's text and split unit tests
Split tests into multiple files. Fixed validator which caused test to fail.
