+++ DoNotExpire/manager/tests/test_validators.py
@@ -1,11 +1,11 @@
 from django.contrib.auth.models import User
 from django.test import TestCase
 
-from DoNotExpire.manager.models import Account, Character
-from .forms import CreateAccountForm
+from DoNotExpire.manager.forms import CreateAccountForm
 
 
-class RegexValidatorsOnModelFormTests(TestCase):
+# to be finished
+class ModelValidatorsTests(TestCase):
     @classmethod
     def setUpTestData(cls) -> None:
         cls.user = User.objects.create_user(
@@ -52,34 +52,14 @@ class RegexValidatorsOnModelFormTests(TestCase):
     def test_letters_only_validator_success(self):
         pass
 
-
-class ManagerModelsTests(TestCase):
-    @classmethod
-    def setUpTestData(cls) -> None:
-        cls.user = User.objects.create_user(
-            'TestUser', email='testuser@tested.com', password='testing')
-        cls.acc = Account.objects.create(
-            name="TestAcc3", profile=cls.user.profile, realm="US West")
-        cls.char1 = Character.objects.create(
-            name="TestChar1", level=99, char_class="Barbarian", acc=cls.acc)
-        cls.char2 = Character.objects.create(
-            name="TestChar2", level=1, char_class="Paladin", acc=cls.acc)
-
-    def test_acc_get_all_characters(self):
+    def test_one_underscore_in_string(self):
         pass
 
-    def test_acc_get_all_characters_count(self):
+    def test_multiple_underscores_in_string(self):
         pass
 
-    def test_char_get_class_image(self):
+    def test_string_starts_with_underscore(self):
         pass
 
-    def test_char_expires_in(self):
+    def test_string_ends_with_underscore(self):
         pass
-
-    def test_equipment_is_created_with_character(self):
-        pass
-
-
-class ManagerViewsTests(TestCase):
-    pass
