+++ DoNotExpire/manager/tests/test_views.py
@@ -1,0 +1,25 @@
+from django.test import TestCase
+
+
+# to do
+class ManagerViewsTests(TestCase):
+    def setUp(self) -> None:
+        pass
+
+    def test_home_is_authenticated(self):
+        pass
+
+    def test_home_not_authenticated(self):
+        pass
+
+    def test_create_character(self):
+        pass
+
+    def test_create_account(self):
+        pass
+
+    def test_delete_account(self):
+        pass
+
+    def test_update_date(self):
+        pass
